Restructure front-end queries and mutations files
