+++ client/src/components/Recommended.js
@@ -3,7 +3,7 @@ import { makeStyles, useTheme } from '@material-ui/core/styles';
 import { Card, CardContent, Typography, IconButton } from '@material-ui/core';
 import { ArrowBack, ArrowForward } from '@material-ui/icons';
 import { gql, useQuery } from '@apollo/client';
-import { QUERY_TUTORIALS } from '../utils/tutorialQuery';
+import { QUERY_TUTORIALS } from '../utils/queries/tutorialQueries';
 
 const useStyles = makeStyles((theme) => ({
   recommendations: {
