--- client/src/utils/categoryQuery.js
@@ -1,10 +1,0 @@
-import { gql } from '@apollo/client';
-
-export const QUERY_CATEGORIES = gql`
-  {
-    categories {
-      _id
-      category
-    }
-  }
-`;
